tests/wpt/tests/html/canvas/offscreen/text/2d.text.drawing.style.textRendering.settings.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.text.drawing.style.textRendering.settings</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.text.drawing.style.textRendering.settings</h1>
<p class="desc">Testing basic functionalities of textRendering in Canvas</p>


<script>
var t = async_test("Testing basic functionalities of textRendering in Canvas");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  // Setting textRendering with correct case.
  _assertSame(ctx.textRendering, "auto", "ctx.textRendering", "\"auto\"");

  ctx.textRendering = "optimizeSpeed";
  _assertSame(ctx.textRendering, "optimizeSpeed", "ctx.textRendering", "\"optimizeSpeed\"");

  ctx.textRendering = "optimizeLegibility";
  _assertSame(ctx.textRendering, "optimizeLegibility", "ctx.textRendering", "\"optimizeLegibility\"");

  ctx.textRendering = "geometricPrecision";
  _assertSame(ctx.textRendering, "geometricPrecision", "ctx.textRendering", "\"geometricPrecision\"");

  ctx.textRendering = "auto";
  _assertSame(ctx.textRendering, "auto", "ctx.textRendering", "\"auto\"");

  // Setting textRendering with incorrect case is ignored.
  ctx.textRendering = "OPtimizeSpeed";
  _assertSame(ctx.textRendering, "auto", "ctx.textRendering", "\"auto\"");

  ctx.textRendering = "OPtimizELEgibility";
  _assertSame(ctx.textRendering, "auto", "ctx.textRendering", "\"auto\"");

  ctx.textRendering = "GeometricPrecision";
  _assertSame(ctx.textRendering, "auto", "ctx.textRendering", "\"auto\"");

  ctx.textRendering = "optimizespeed";
  _assertSame(ctx.textRendering, "auto", "ctx.textRendering", "\"auto\"");

  ctx.textRendering = "optimizelegibility";
  _assertSame(ctx.textRendering, "auto", "ctx.textRendering", "\"auto\"");

  ctx.textRendering = "geometricprecision";
  _assertSame(ctx.textRendering, "auto", "ctx.textRendering", "\"auto\"");

  ctx.textRendering = "optimizeLegibility";
  _assertSame(ctx.textRendering, "optimizeLegibility", "ctx.textRendering", "\"optimizeLegibility\"");

  ctx.textRendering = "AUTO";
  _assertSame(ctx.textRendering, "optimizeLegibility", "ctx.textRendering", "\"optimizeLegibility\"");

  ctx.textRendering = "Auto";
  _assertSame(ctx.textRendering, "optimizeLegibility", "ctx.textRendering", "\"optimizeLegibility\"");

  // Setting textRendering with non-existing font variant.
  ctx.textRendering = "abcd";
  _assertSame(ctx.textRendering, "optimizeLegibility", "ctx.textRendering", "\"optimizeLegibility\"");

  ctx.textRendering = "normal";
  _assertSame(ctx.textRendering, "optimizeLegibility", "ctx.textRendering", "\"optimizeLegibility\"");

  ctx.textRendering = "";
  _assertSame(ctx.textRendering, "optimizeLegibility", "ctx.textRendering", "\"optimizeLegibility\"");

  ctx.textRendering = "auto";
  _assertSame(ctx.textRendering, "auto", "ctx.textRendering", "\"auto\"");
  t.done();

});
</script>
